requirements.txt
mlflow==2.21.3
mlflow_skinny==2.21.3
pydantic==2.11.4
pytest==8.3.5
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
torch==2.6.0
Wrapper==1.1.0b1
